+++ pmd-ui/src/main/java/net/sourceforge/pmd/util/fxdesigner/Designer.java
@@ -10,10 +10,13 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
+import org.apache.commons.lang3.ArrayUtils;
+
 import net.sourceforge.pmd.PMDVersion;
 import net.sourceforge.pmd.util.fxdesigner.app.DesignerRoot;
 import net.sourceforge.pmd.util.fxdesigner.util.DesignerUtil;
 
+import com.sun.javafx.fxml.builder.ProxyBuilder;
 import javafx.application.Application;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXMLLoader;
@@ -31,23 +34,30 @@ import javafx.stage.Stage;
  */
 public class Designer extends Application {
 
-    private boolean parseParameters(Parameters params) {
+    private boolean isDeveloperMode;
+    private long initStartTimeMillis;
+
+    public Designer() {
+        initStartTimeMillis = System.currentTimeMillis();
+    }
+
+    private void parseParameters(Parameters params) {
         List<String> raw = params.getRaw();
-        if (!raw.contains("-v")
-            && !raw.contains("--verbose")) {
-            // error output is disabled by default
-            
-            System.err.close();
-            return false;
+        // error output is disabled by default
+        if (raw.contains("-v") || raw.contains("--verbose")) {
+            isDeveloperMode = true;
         }
-        return true;
     }
 
 
     @Override
     public void start(Stage stage) throws IOException {
-        boolean isDeveloperMode = parseParameters(getParameters());
+        parseParameters(getParameters());
 
+        stage.setTitle("PMD Rule Designer (v " + PMDVersion.VERSION + ')');
+        setIcons(stage);
+
+        System.out.print(stage.getTitle() + " initializing... ");
 
         FXMLLoader loader = new FXMLLoader(DesignerUtil.getFxml("designer.fxml"));
 
@@ -57,6 +67,23 @@ public class Designer extends Application {
         NodeInfoPanelController nodeInfoPanelController = new NodeInfoPanelController(mainController);
         XPathPanelController xpathPanelController = new XPathPanelController(mainController);
         SourceEditorController sourceEditorController = new SourceEditorController(mainController);
+
+        loader.setBuilderFactory(type -> {
+
+            boolean needsRoot = Arrays.stream(type.getConstructors()).anyMatch(it -> ArrayUtils.contains(it.getParameterTypes(), DesignerRoot.class));
+
+            if (needsRoot) {
+                // Controls that need the DesignerRoot can declare a constructor
+                // with a parameter w/ signature @NamedArg("designerRoot") DesignerRoot
+                // to be injected with the relevant instance of the app.
+                // TODO Not everything has been refactored to use this mechanism for now
+                ProxyBuilder<Object> builder = new ProxyBuilder<>(type);
+                builder.put("designerRoot", owner);
+                return builder;
+            } else {
+                return null; //use default
+            }
+        });
 
         loader.setControllerFactory(type -> {
             if (type == MainDesignerController.class) {
@@ -83,11 +110,22 @@ public class Designer extends Application {
         Parent root = loader.load();
         Scene scene = new Scene(root);
 
-        stage.setTitle("PMD Rule Designer (v " + PMDVersion.VERSION + ')');
-        setIcons(stage);
-
         stage.setScene(scene);
+
+        if (owner.isDeveloperMode()) {
+            // only close after initialization succeeded.
+            // but before stage.show to reduce unwanted noise
+            System.err.close();
+        }
+
         stage.show();
+
+        long initTime = System.currentTimeMillis() - initStartTimeMillis;
+
+        System.out.println("done in " + initTime + "ms.");
+        if (!owner.isDeveloperMode()) {
+            System.out.println("Run with --verbose parameter to enable error output.");
+        }
     }