+++ pmd-ui/src/main/java/net/sourceforge/pmd/util/fxdesigner/SourceEditorController.java
@@ -5,6 +5,11 @@
 package net.sourceforge.pmd.util.fxdesigner;
 
 import static java.util.Collections.emptyList;
+import static net.sourceforge.pmd.util.fxdesigner.util.DesignerUtil.defaultLanguageVersion;
+import static net.sourceforge.pmd.util.fxdesigner.util.DesignerUtil.getSupportedLanguageVersions;
+import static net.sourceforge.pmd.util.fxdesigner.util.DesignerUtil.mapToggleGroupToUserData;
+import static net.sourceforge.pmd.util.fxdesigner.util.DesignerUtil.rewire;
+import static net.sourceforge.pmd.util.fxdesigner.util.DesignerUtil.sanitizeExceptionMessage;
 
 import java.io.File;
 import java.io.IOException;
@@ -78,16 +83,11 @@ public class SourceEditorController extends AbstractController<MainDesignerContr
     public SourceEditorController(MainDesignerController mainController) {
         super(mainController);
         astManager = new ASTManager(mainController.getDesignerRoot());
-
     }
 
 
     @Override
     protected void beforeParentInit() {
-
-        astTreeView.setDesignerRoot(getDesignerRoot());
-        nodeEditionCodeArea.setDesignerRoot(getDesignerRoot());
-
         initializeLanguageSelector(); // languageVersionProperty() must be initialized
 
         languageVersionProperty().values()
@@ -117,7 +117,7 @@ public class SourceEditorController extends AbstractController<MainDesignerContr
 
     @Override
     protected void afterParentInit() {
-        DesignerUtil.rewire(astManager.languageVersionProperty(), languageVersionUIProperty);
+        rewire(astManager.languageVersionProperty(), languageVersionUIProperty);
         nodeEditionCodeArea.moveCaret(0, 0);
     }
 
@@ -126,7 +126,7 @@ public class SourceEditorController extends AbstractController<MainDesignerContr
 
         ToggleGroup languageToggleGroup = new ToggleGroup();
 
-        DesignerUtil.getSupportedLanguageVersions()
+        getSupportedLanguageVersions()
                     .stream()
                     .sorted(LanguageVersion::compareTo)
                     .map(lv -> {
@@ -139,9 +139,9 @@ public class SourceEditorController extends AbstractController<MainDesignerContr
                         languageSelectionMenuButton.getItems().add(item);
                     });
 
-        languageVersionUIProperty = DesignerUtil.mapToggleGroupToUserData(languageToggleGroup, DesignerUtil::defaultLanguageVersion);
+        languageVersionUIProperty = mapToggleGroupToUserData(languageToggleGroup, DesignerUtil::defaultLanguageVersion);
         // this will be overwritten by property restore if needed
-        languageVersionUIProperty.setValue(DesignerUtil.defaultLanguageVersion());
+        languageVersionUIProperty.setValue(defaultLanguageVersion());
     }
 
     /**
@@ -160,7 +160,7 @@ public class SourceEditorController extends AbstractController<MainDesignerContr
         try {
             current = astManager.updateIfChanged(source, auxclasspathClassLoader.getValue());
         } catch (ParseAbortedException e) {
-            astTitledPane.setTitle("Abstract syntax tree (error)");
+            editorTitledPane.errorMessageProperty().setValue(sanitizeExceptionMessage(e));
             return Optional.empty();
         }
 
@@ -176,7 +176,7 @@ public class SourceEditorController extends AbstractController<MainDesignerContr
 
     private void setUpToDateCompilationUnit(Node node) {
         parent.invalidateAst();
-        astTitledPane.setTitle("Abstract syntax tree");
+        editorTitledPane.errorMessageProperty().setValue("");
         ASTTreeItem root = ASTTreeItem.getRoot(node);
         astTreeView.setRoot(root);
     }