+++ pmd-ui/src/main/java/net/sourceforge/pmd/util/fxdesigner/XPathPanelController.java
@@ -5,6 +5,8 @@
 package net.sourceforge.pmd.util.fxdesigner;
 
 
+import static net.sourceforge.pmd.util.fxdesigner.util.DesignerUtil.sanitizeExceptionMessage;
+
 import java.io.IOException;
 import java.time.Duration;
 import java.util.ArrayList;
@@ -13,12 +15,14 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Supplier;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.StringUtils;
 import org.controlsfx.validation.ValidationSupport;
 import org.controlsfx.validation.Validator;
-import org.reactfx.EventStream;
+import org.kordamp.ikonli.javafx.FontIcon;
 import org.reactfx.EventStreams;
 import org.reactfx.SuspendableEventStream;
 import org.reactfx.collection.LiveArrayList;
@@ -56,6 +60,8 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.ContextMenu;
+import javafx.scene.control.Hyperlink;
+import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 import javafx.scene.control.MenuButton;
 import javafx.scene.control.MenuItem;
@@ -64,6 +70,9 @@ import javafx.scene.control.TextArea;
 import javafx.scene.control.ToggleGroup;
 import javafx.scene.input.MouseButton;
 import javafx.scene.input.MouseEvent;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.VBox;
+import javafx.scene.text.Text;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
@@ -79,6 +88,7 @@ import javafx.stage.StageStyle;
  */
 public class XPathPanelController extends AbstractController<MainDesignerController> implements NodeSelectionSource {
 
+    private static final String NO_MATCH_MESSAGE = "No match in text";
     private static final Duration XPATH_REFRESH_DELAY = Duration.ofMillis(100);
     private final XPathEvaluator xpathEvaluator = new XPathEvaluator();
     private final ObservableXPathRuleBuilder ruleBuilder = new ObservableXPathRuleBuilder();
@@ -125,17 +135,18 @@ public class XPathPanelController extends AbstractController<MainDesignerControl
 
         exportXpathToRuleButton.setOnAction(e -> showExportXPathToRuleWizard());
 
-        xpathExpressionArea.richChanges()
-                           .filter(t -> !t.isIdentity())
-                           .successionEnds(XPATH_REFRESH_DELAY)
-                           // Reevaluate XPath anytime the expression or the XPath version changes
-                           .or(xpathVersionProperty().changes())
-                           .subscribe(tick -> parent.refreshXPathResults());
+        getRuleBuilder().modificationsTicks()
+                        .successionEnds(XPATH_REFRESH_DELAY)
+                        .subscribe(tick -> parent.refreshXPathResults());
 
         selectionEvents = EventStreams.valuesOf(xpathResultListView.getSelectionModel().selectedItemProperty()).suppressible();
 
-        initNodeSelectionHandling();
+        initNodeSelectionHandling(getDesignerRoot(),
+                                  selectionEvents.filter(Objects::nonNull).map(TextAwareNodeWrapper::getNode), false);
+
+        violationsTitledPane.titleProperty().bind(currentResults.map(List::size).map(n -> "Matched nodes (" + n + ")"));
     }
+
 
 
     @Override
@@ -223,14 +234,6 @@ public class XPathPanelController extends AbstractController<MainDesignerControl
         });
     }
 
-
-    @Override
-    public EventStream<Node> getSelectionEvents() {
-        return selectionEvents.filter(Objects::nonNull)
-                              .map(TextAwareNodeWrapper::getNode);
-    }
-
-
     @Override
     public void setFocusNode(Node node) {
         Optional<TextAwareNodeWrapper> firstResult = xpathResultListView.getItems().stream()
@@ -260,7 +263,7 @@ public class XPathPanelController extends AbstractController<MainDesignerControl
         try {
             String xpath = getXpathExpression();
             if (StringUtils.isBlank(xpath)) {
-                invalidateResults(false);
+                updateResults(false, false, Collections.emptyList(), "Type an XPath expression to show results");
                 return;
             }
 
@@ -270,13 +273,12 @@ public class XPathPanelController extends AbstractController<MainDesignerControl
                                                                                  getXpathVersion(),
                                                                                  xpath,
                                                                                  ruleBuilder.getRuleProperties()));
-            xpathResultListView.setItems(results.stream().map(parent::wrapNode).collect(Collectors.toCollection(LiveArrayList::new)));
-            this.currentResults.setValue(results);
-            violationsTitledPane.setTitle("Matched nodes (" + results.size() + ")");
+
+            updateResults(false, false, results, NO_MATCH_MESSAGE);
             // Notify that everything went OK so we can avoid logging very recent exceptions
             raiseParsableXPathFlag();
         } catch (XPathEvaluationException e) {
-            invalidateResults(true);
+            updateResults(true, false, Collections.emptyList(), sanitizeExceptionMessage(e));
             logUserException(e, Category.XPATH_EVALUATION_EXCEPTION);
         }
 
@@ -288,10 +290,12 @@ public class XPathPanelController extends AbstractController<MainDesignerControl
     }
 
 
-    public void invalidateResults(boolean error) {
-        xpathResultListView.getItems().clear();
-        this.currentResults.setValue(Collections.emptyList());
-        violationsTitledPane.setTitle("Matched nodes" + (error ? "\t(error)" : ""));
+    /**
+     * Called by the rest of the app.
+     */
+    public void invalidateResultsExternal(boolean error) {
+        String placeholder = error ? "Compilation unit is invalid" : NO_MATCH_MESSAGE;
+        updateResults(false, true, Collections.emptyList(), placeholder);
     }
 
 
@@ -364,4 +368,53 @@ public class XPathPanelController extends AbstractController<MainDesignerControl
     }
 
 
+    private static final Pattern JAXEN_MISSING_PROPERTY_EXTRACTOR = Pattern.compile("Variable (\\w+)");
+    private static final Pattern SAXON_MISSING_PROPERTY_EXTRACTOR = Pattern.compile("Undeclared variable in XPath expression: \\$(\\w+)");
+
+
+    private void updateResults(boolean xpathError,
+                               boolean otherError,
+                               List<Node> results,
+                               String emptyResultsPlaceholder) {
+
+        javafx.scene.Node emptyLabel = xpathError || otherError
+                                       ? getErrorPlaceholder(emptyResultsPlaceholder)
+                                       : new Label(emptyResultsPlaceholder);
+
+        xpathResultListView.setPlaceholder(emptyLabel);
+
+        xpathResultListView.setItems(results.stream().map(parent::wrapNode).collect(Collectors.toCollection(LiveArrayList::new)));
+        this.currentResults.setValue(results);
+        // only show the error label here when it's an xpath error
+        expressionTitledPane.errorMessageProperty().setValue(xpathError ? emptyResultsPlaceholder : "");
+    }
+
+
+    private javafx.scene.Node getErrorPlaceholder(String message) {
+        VBox vbox = getMissingPropertyName(message).map(
+            name -> {
+                Hyperlink hyperlink = new Hyperlink("Add property");
+                hyperlink.setOnAction(e -> propertyTableView.onAddPropertyClicked(name));
+                return new VBox(new Text("Undeclared property in XPath expression: $" + name), hyperlink);
+            }
+        ).orElse(new VBox(new Text(message)));
+
+        HBox hbox = new HBox(new FontIcon("fas-exclamation-triangle"), vbox);
+        hbox.setSpacing(10);
+        hbox.maxHeightProperty().bind(hbox.heightProperty());
+        hbox.maxWidthProperty().bind(vbox.widthProperty());
+        hbox.setFillHeight(false);
+        vbox.setFillWidth(false);
+        return hbox;
+    }
+
+
+    private Optional<String> getMissingPropertyName(String errorMessage) {
+        Pattern nameExtractor = XPathRuleQuery.XPATH_1_0.equals(getXpathVersion())
+                                ? JAXEN_MISSING_PROPERTY_EXTRACTOR
+                                : SAXON_MISSING_PROPERTY_EXTRACTOR;
+
+        Matcher matcher = nameExtractor.matcher(errorMessage);
+        return matcher.matches() ? Optional.of(matcher.group(1)) : Optional.empty();
+    }
 }